Improving the code design
